Revert "update test"
This reverts commit 411f1d7f384e9497a38a3d7b0ec1922d9953b671.

+++ backend/internals/services/profile_service_test.go
@@ -80,7 +80,7 @@ func (suite *ProfileTestSuit) TestGetTotalGemsWhenSuccess() {
 	mockTotalGems := 100
 
 	// mock user repo
-	mockUserRepo.EXPECT().GetTotalGemsByUserID(mockUserID).Return(uint64(mockTotalGems), nil)
+	mockUserRepo.EXPECT().GetTotalGemsByUserID(mockUserID).Return(mockTotalGems, nil)
 
 	// Test
 	underTest := services.NewProfileService(mockUserRepo)
